+++ src/adapters/async_read.rs
@@ -100,7 +100,6 @@ where
 
 #[cfg(test)]
 mod tests {
-    #[allow(unused_imports)]
     use super::*;
 
     #[tokio::test]