+++ src/adapters/mod.rs
@@ -3,5 +3,7 @@
 //! [`Body`]: crate::Body
 
 mod async_read;
+mod read;
 
 pub use async_read::AsyncReadBody;
+pub use read::ReadBody;